requirements.txt
docopt
passlib
aiohttp
autobahn
pyasn1
aiopg
yoyo-migrations
autobahn_sync
sqlalchemy
